refactor(Repeater): Split SelectionModel/SelectionNode/SelectionTreeHelper per winui-port
Splits the selection types per the winui-port file layout (.cs declaration,
.h.mux.cs fields, .mux.cs implementation). Replaces the legacy SelectionModel.Header.cs
/ SelectionNode.Header.cs naming. SelectedItemInfo struct moves to
SelectionModel.h.mux.cs. Bodies are 1:1 with WinUI commit 4b206bce3, MUX_ASSERT
reinstated where the previous port used Debug.Assert.

Tracks issue #22908.
